Desktop/Proyecto sst/proyecto_final_sst/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4